+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Added function utils.addFacetedCriteria to ease applying a faceted conf xml
   that adds extra faceted criteria to an existing dashboard.
   [gbastien]
+- Improved template evaluate method to avoid getting collection and criterias if not necessary
+  [sgeulette]
 
 0.8 (2019-05-16)
 ----------------